refactor(web): extract list command functions into separate module
Move unwrapList, wrapBlocksInList, switchListType, and helper functions (findParentList, collectListItems, blockRange, hasTaskListAncestor) from MilkdownEditor.tsx into a dedicated listCommands.ts module.

This is pure extraction with no behavioral changes. The functions remain identical — they're now importable for direct testing without a browser.
